# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet EDR API
  version: "0.1.0"
  description: |
    HTTP API for Fleet EDR. Three audience tiers:

    - **Agents** (macOS endpoints) authenticate with a per-host bearer
      token obtained at enrollment. They call `POST /api/enroll`,
      `POST /api/events`, and the host-scoped `/commands` endpoints.
    - **Browsers** (admin UI) authenticate with an `edr_session` cookie
      (HttpOnly, Secure, SameSite=Lax) plus an `X-CSRF-Token` header on
      unsafe methods. They call the session-protected endpoints.
    - **Operators** hit `/livez`, `/readyz`, `/health` with no auth.

    For prose context see
    [docs/api.md](https://github.com/getvictor/fleet-edr/blob/main/docs/api.md).
    For a human-readable host/agent install guide see
    [docs/install-agent-manual.md](https://github.com/getvictor/fleet-edr/blob/main/docs/install-agent-manual.md).
  license:
    name: Fleet EDR License
    url: https://github.com/getvictor/fleet-edr/blob/main/LICENSE
servers:
  - url: https://edr.example.com
    description: Your EDR server
...
paths:
...
  /api/events:
    post:
      tags: [Agent]
      summary: Ingest a batch of endpoint security events
      description: |
        Agents POST a JSON array of event envelopes. Each event's
        `host_id` must match the token's authenticated host_id;
        mismatches reject the whole batch. The body is capped at 10 MB.
        The server returns the count of events persisted.
      security:
        - HostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/EventEnvelope"
      responses:
        "200":
          description: Events persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid JSON, missing required fields, or host_id mismatch
        "401":
          description: Missing or invalid host token
  /api/session:
    post:
      tags: [Session]
      summary: Log in with email + password
      description: |
        Validates credentials, creates a session row, sets `edr_session`
        cookie, returns the user + CSRF token. Rate-limited per-IP at
        `EDR_LOGIN_RATE_PER_MIN` (default 6).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          headers:
            Set-Cookie:
              description: Session cookie (HttpOnly, Secure, SameSite=Lax)
              schema:
                type: string
                example: edr_session=<opaque>; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Malformed request
        "401":
          description: Wrong email or password
        "429":
          description: Rate limited
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: {type: integer}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [Session]
      summary: Get the current session
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not logged in
    delete:
      tags: [Session]
      summary: Log out
      description: Clears the session row and expires the cookie.
      security:
        - SessionCookie: []
          CSRF: []
      responses:
        "204":
          description: Logged out
  /api/hosts:
    get:
      tags: [Hosts]
      summary: List hosts
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Host summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostSummary"
  /api/hosts/{host_id}/tree:
    get:
      tags: [Hosts]
      summary: Process tree for a host
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: from
          in: query
          schema: {type: integer, format: int64}
          description: Lower bound, ns since epoch. Default `now - 1h`.
        - name: to
          in: query
          schema: {type: integer, format: int64}
          description: Upper bound, ns since epoch. Default `now`.
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 5000, default: 2000}
      responses:
        "200":
          description: Forest of process roots
          content:
            application/json:
              schema:
                type: object
                properties:
                  roots:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProcessNode"
  /api/hosts/{host_id}/processes/{pid}:
    get:
      tags: [Hosts]
      summary: Process detail
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: pid
          in: path
          required: true
          schema: {type: integer}
        - name: at
          in: query
          schema: {type: integer, format: int64}
          description: Point-in-time ns since epoch. Default `now`.
      responses:
        "200":
          description: Process detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: No process found at the given coordinates
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      security:
        - SessionCookie: []
      parameters:
        - name: host_id
          in: query
          schema: {type: string}
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: process_id
          in: query
          schema: {type: integer, format: int64}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, default: 100}
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /api/alerts/{id}:
    get:
      tags: [Alerts]
      summary: Alert detail (with event IDs)
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Alert with linked events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertDetail"
        "404":
          description: Alert not found
    put:
      tags: [Alerts]
      summary: Update alert status
      security:
        - SessionCookie: []
          CSRF: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [open, acknowledged, resolved]
      responses:
        "204":
          description: Status updated
        "400":
          description: Invalid status value
        "404":
          description: Alert not found
  /api/commands:
    get:
      tags: [Commands]
      summary: List commands for the authenticated host
      description: |
        Host-token-only. The `host_id` is always the authenticated host's
        id; any query parameter is ignored. Polling this endpoint also
        refreshes the host's `last_seen` timestamp.
      security:
        - HostToken: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, acked, completed, failed]
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
    post:
      tags: [Commands]
      summary: Enqueue a command for a host (admin)
      security:
        - SessionCookie: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id, command_type]
              properties:
                host_id: {type: string}
                command_type: {type: string}
                payload: {type: object, additionalProperties: true}
      responses:
        "201":
          description: Command created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: integer, format: int64}
  /api/commands/{id}:
    get:
      tags: [Commands]
      summary: Get a single command (admin)
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        "404":
          description: Not found
    put:
      tags: [Commands]
      summary: Update command status (agent)
      description: |
        Agents call this with `status=acked` once they pick up a command
        and with `status=completed` or `status=failed` once they finish.
        The handler enforces that the command belongs to the authenticated
        host — agent A cannot touch agent B's commands.
      security:
        - HostToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [acked, completed, failed]
                result:
                  type: object
                  description: Command-type-specific result blob, free-form JSON
      responses:
        "204":
          description: Status updated
        "400":
          description: Invalid status
        "404":
          description: Command not found (or not owned by authenticated host)
  /api/enrollments:
    get:
      tags: [Admin]
      summary: List enrollments
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Enrollment rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
  /api/enrollments/{host_id}/revoke:
    post:
      tags: [Admin]
      summary: Revoke a host enrollment
      description: |
        After revocation the host's token stops working on all
        host-token endpoints. A revoked host cannot re-enroll without
        admin approval (reusing the same hardware UUID still returns
        403).
      security:
        - SessionCookie: []
          CSRF: []
      parameters:
        - $ref: "#/components/parameters/HostID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
                actor: {type: string}
      responses:
        "204":
          description: Revoked
        "404":
          description: Host not found
  /api/policy:
    get:
      tags: [Admin]
      summary: Current blocklist policy
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    put:
      tags: [Admin]
      summary: Replace blocklist policy
      description: |
        Replaces the entire policy atomically and fans out a
        `policy.update` command to every active host. The response
        reports how many hosts received the fan-out; failed hosts will
        pick it up on their next poll.
      security:
        - SessionCookie: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paths:
                  type: array
                  items: {type: string}
                hashes:
                  type: array
                  items: {type: string}
                reason: {type: string}
                actor: {type: string}
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyUpdateResult"
  /api/attack-coverage:
    get:
      tags: [Admin]
      summary: MITRE ATT&CK coverage (Navigator layer)
      description: |
        Returns a MITRE ATT&CK Navigator layer document enumerating the
        techniques covered by currently-registered detection rules. Drop
        the response directly into
        https://mitre-attack.github.io/attack-navigator/ to visualise
        coverage as a heatmap. Output is deterministic (sorted technique
        IDs + sorted, dedup'd rule ID lists), so responses are
        byte-identical across requests and safe to ETag.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Navigator layer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigatorLayer"
  /api/rules:
    get:
      tags: [Admin]
      summary: Detection rule documentation
      description: |
        Returns the structured per-rule documentation for every registered
        detection rule, in the engine's registration order. The same data
        feeds the UI's `/ui/rules/<id>` rule-detail page and the
        deterministic markdown reference at `docs/detection-rules.md`
        (generated by `tools/gen-rule-docs`).
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Per-rule documentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RulesResponse"
components:
  securitySchemes:
    HostToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Per-host bearer token issued by `POST /api/enroll`. Pass as
        `Authorization: Bearer <host_token>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: edr_session
      description: |
        HttpOnly cookie set by `POST /api/session`. The browser
        attaches it automatically.
    CSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on unsafe methods (`POST`, `PUT`, `DELETE`) for
        session-cookie-authenticated endpoints. Value comes from the
        login response's `csrf_token` field.
  parameters:
    HostID:
      name: host_id
      in: path
      required: true
      description: Host identifier (UUID-shaped opaque string)
      schema: {type: string}
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    LoginResponse:
      type: object
      required: [user, csrf_token]
      properties:
        user:
          type: object
          required: [id, email]
          properties:
            id: {type: integer, format: int64}
            email: {type: string, format: email}
        csrf_token: {type: string}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: {type: string}
    EventEnvelope:
      type: object
      required: [event_id, host_id, timestamp_ns, event_type, payload]
      properties:
        event_id:
          type: string
          description: Opaque unique event identifier (the agent emits UUIDs; the server treats it as an opaque string).
        host_id: {type: string}
        timestamp_ns: {type: integer, format: int64}
        event_type:
          type: string
          enum: [exec, fork, exit, open, network_connect, dns_query]
        payload:
          type: object
          description: Event-type-specific body; see schema/events.json in the repo
    IngestResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: integer
          description: Number of events successfully persisted.
    HostSummary:
      type: object
      properties:
        host_id: {type: string}
        event_count: {type: integer, format: int64}
        last_seen_ns: {type: integer, format: int64}
    Process:
      type: object
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        pid: {type: integer}
        ppid: {type: integer}
        path: {type: string}
        args: {type: array, items: {type: string}}
        uid: {type: integer}
        gid: {type: integer}
        code_signing:
          type: object
          properties:
            team_id: {type: string}
            signing_id: {type: string}
            flags: {type: integer}
            is_platform_binary: {type: boolean}
        sha256: {type: string}
        fork_time_ns: {type: integer, format: int64}
        exec_time_ns: {type: integer, format: int64}
        exit_time_ns: {type: integer, format: int64}
        exit_code: {type: integer}
    ProcessNode:
      allOf:
        - $ref: "#/components/schemas/Process"
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: "#/components/schemas/ProcessNode"
    Alert:
      type: object
      required: [id, host_id, rule_id, severity, title, process_id, status, created_at, updated_at]
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        rule_id: {type: string}
        severity: {type: string, enum: [low, medium, high, critical]}
        title: {type: string}
        description: {type: string}
        process_id: {type: integer, format: int64}
        status: {type: string, enum: [open, acknowledged, resolved]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        resolved_at: {type: string, format: date-time}
    AlertDetail:
      allOf:
        - $ref: "#/components/schemas/Alert"
        - type: object
          properties:
            event_ids:
              type: array
              items: {type: string, format: uuid}
...
    Policy:
      type: object
      properties:
        version: {type: integer, format: int64}
        blocklist:
          type: object
          properties:
            paths:
              type: array
              items: {type: string}
            hashes:
              type: array
              items: {type: string}
...